feat: Add PJE and PROJUDI modules for process automation
- Introduced PJE_AM and PROJUDI_AM classes for handling respective systems.
- Implemented configuration properties in properties.py for element management.
- Developed search functionalities for PJE, ESAJ, and PROJUDI systems.
- Enhanced Excel template creation with MakeTemplates and Listas classes.
- Updated search controller to manage process searches and captcha challenges.
- Refactored element imports to align with new module structure.
- Added comprehensive docstrings in Google style for better documentation.
