+++ celery_app/addons/auth/elaw.py
@@ -6,7 +6,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 
-from crawjud.addons.auth import AuthController
+from celery_app.addons.auth import AuthController
 
 
 class ElawAuth(AuthController):
