+++ celery_app/addons/auth/pje.py
@@ -7,7 +7,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.wait import WebDriverWait
 
-from crawjud.addons.auth import AuthController
+from celery_app.addons.auth import AuthController
 from utils.webdriver import DriverBot
