+++ celery_app/addons/elements/__init__.py
@@ -20,10 +20,10 @@ from __future__ import annotations
 from importlib import import_module
 from typing import AnyStr, Self
 
-from crawjud.addons.elements.elaw import ELAW_AME
-from crawjud.addons.elements.esaj import ESAJ_AM
-from crawjud.addons.elements.pje import PJE_AM
-from crawjud.addons.elements.projudi import PROJUDI_AM
+from celery_app.addons.elements.elaw import ELAW_AME
+from celery_app.addons.elements.esaj import ESAJ_AM
+from celery_app.addons.elements.pje import PJE_AM
+from celery_app.addons.elements.projudi import PROJUDI_AM
 
 
 class ElementsBot:
