+++ celery_app/addons/elements/caixa.py
@@ -1,0 +1,17 @@
+"""Module for handling Caixa AM elements.
+
+Configure and manage elements for Caixa-related processes.
+"""
+
+
+class CAIXA_AM:  # noqa: N801
+    """Represent Caixa AM elements.
+
+    Provide methods and attributes to operate Caixa AM related functionality.
+
+    Attributes:
+        # Define attributes here as needed.
+
+    """
+
+    # Elementos Caixa
