+++ celery_app/addons/search/esaj.py
@@ -5,14 +5,14 @@ from __future__ import annotations
 from contextlib import suppress
 from time import sleep
 
+from crawjud.common.exceptions.bot import ExecutionError
 from selenium.common.exceptions import NoSuchElementException, TimeoutException
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.wait import WebDriverWait
 
-from crawjud.addons.search import SearchController
-from crawjud.common.exceptions.bot import ExecutionError
+from celery_app.addons.search import SearchController
 
 
 class EsajSearch(SearchController):
