+++ celery_app/addons/search/projudi.py
@@ -6,13 +6,13 @@ from time import sleep
 from typing import TYPE_CHECKING
 
 import pytz
+from crawjud.common.exceptions.bot import ExecutionError
 from selenium.common.exceptions import NoSuchElementException, TimeoutException
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.wait import WebDriverWait
 
-from crawjud.addons.search import SearchController
-from crawjud.common.exceptions.bot import ExecutionError
+from celery_app.addons.search import SearchController
 
 if TYPE_CHECKING:
     from selenium.webdriver.remote.webelement import WebElement
