--- crawjud/addons/elements/calculadoras.py
@@ -1,20 +1,0 @@
-"""Module: calculadoras.
-
-Provide TJDFT elements for legal processing tasks.
-
-Classes:
-    TJDFT: Represent TJDFT-related elements.
-"""
-
-
-class TJDFT:
-    """Represent TJDFT elements.
-
-    Encapsulate elements and functionality specific to the TJDFT system.
-
-    Attributes:
-        # Define attributes here as needed.
-
-    """
-
-    # ELEMENTOS TJDFT
